pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lumen"
version = "3.0.0"
description = "LLM-powered systematic review and meta-analysis pipeline"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "litellm>=1.50",
    "pydantic>=2.0",
    "pdfplumber>=0.11",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "matplotlib>=3.8",
    "scikit-learn>=1.4",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
lumen = "lumen.interface.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
